+++ tests/testthat/test_flags.R
@@ -17,7 +17,11 @@ test_that("read.glider.seaexplorer.realtime flag names", {
 
 test_that("read.glider.seaexplorer.delayed flag names", {
     directory <- system.file("extdata/sea_explorer/delayed_raw", package = "oceglider")
-    expect_silent(g <- read.glider.seaexplorer.delayed(directory = directory, progressBar = FALSE))
+    expect_silent(g <- read.glider.seaexplorer.raw(
+        directory = directory,
+        pattern = "pld1.raw",
+        progressBar = FALSE
+    ))
     expect_equal(
         g@metadata$flagScheme,
         list(
@@ -31,7 +35,7 @@ test_that("read.glider.seaexplorer.delayed flag names", {
 test_that("read.glider.seaexplorer.realtime flag setting and handling", {
     # This is based on the example given by ?"handleFlags,glider-method"
     directory <- system.file("extdata/sea_explorer/realtime_raw", package = "oceglider")
-    expect_silent(g <- read.glider.seaexplorer.realtime(directory))
+    expect_silent(g <- read.glider.seaexplorer.realtime(directory, progressBar = FALSE))
     # NOTE: this test was more hard-wired before issue40, e.g. it
     # demanded that the number of data read be 2784, but that number
     # became 3435 when read.glider.seaexplorer.delayed() was changed to
@@ -41,7 +45,7 @@ test_that("read.glider.seaexplorer.realtime flag setting and handling", {
     lowSalinity <- which(g[["salinity"]] < 31)
     g2 <- setGliderFlags(g, "salinity", g[["salinity"]] < 31, 3)
     expect_true(all(g2[["salinityFlag"]][lowSalinity] == 3))
-    g3 <- handleGliderFlags(g2)# , c(3, 4, 9)) # use default action, which is "NA"
+    g3 <- handleGliderFlags(g2) # , c(3, 4, 9)) # use default action, which is "NA"
     expect_true(all(g3[["salinityFlag"]][lowSalinity] == 3))
     expect_true(all(is.na(g3[["salinity"]][lowSalinity])))
 })
@@ -49,7 +53,9 @@ test_that("read.glider.seaexplorer.realtime flag setting and handling", {
 test_that("read.glider.seaexplorer.delayed flag setting and handling", {
     # This is based on the example given by ?"handleFlags,glider-method"
     directory <- system.file("extdata/sea_explorer/delayed_raw", package = "oceglider")
-    expect_silent(g <- read.glider.seaexplorer.delayed(directory, progressBar = FALSE))
+    expect_silent(g <- read.glider.seaexplorer.raw(directory, "pld1.raw",
+        progressBar = FALSE
+    ))
     # NOTE: this test was more hard-wired before issue40, e.g. it
     # demanded that the number of data read be 2784, but that number
     # became 3435 when read.glider.seaexplorer.delayed() was changed to